+++ src/Senders/MeteorClientHttpSender.js
@@ -13,15 +13,17 @@ import NullFn from "../NullFn";
  * @extends SenderBase
  */
 const MeteorClientHttpSender = class extends SenderBase {
-  // noinspection JSClassNamingConvention
   /**
    * @constructor
    *
+   * @param {ProcessorBase[]} processors
+   *   Processors to be applied by this sender instead of globally.
    * @param {String} loggerUrl
    *   The absolute URL of the logger server. Usually /logger on the Meteor app.
    */
-  constructor(loggerUrl) {
-    super();
+  constructor(processors = [], loggerUrl) {
+    super(processors);
+
     if (typeof Meteor === "undefined" || !Meteor.isClient) {
       throw new Error("MeteorClientHttpSender is only meant for Meteor client side.");
     }
@@ -37,9 +39,10 @@ const MeteorClientHttpSender = class extends SenderBase {
   }
 
   send(level, message, context) {
+    const processedContext = super.send(level, message, context);
     let data = { level, message };
-    if (typeof context !== "undefined") {
-      data.context = context;
+    if (typeof processedContext !== "undefined") {
+      data.context = processedContext;
     }
 
     let options = {
@@ -47,6 +50,7 @@ const MeteorClientHttpSender = class extends SenderBase {
       headers: this.requestHeaders,
     };
     this.http.post(this.loggerUrl, options, NullFn);
+    return processedContext;
   }
 };