+++ src/ServerLogger.js
@@ -12,6 +12,8 @@ import process from "process";
  *
  * Its main method is log(level, message, context).
  *
+ * @extends Logger
+ *
  * @see ServerLogger.log
  */
 class ServerLogger extends Logger {
@@ -23,13 +25,15 @@ class ServerLogger extends Logger {
    *   A logging strategy instance.
    * @param {WebApp} webapp
    *   The Meteor WebApp service.
+   * @param {Processor[]} processors
+   *   An array of processor instances.
    * @param {Object} parameters
    * - logRequestHeaders: add request headers to the log context. Defaults to true.
    * - servePath: the path on which to expose the logger endpoint. Defaults to "/logger".
    * - enableMethod: enable the filog:log method or not. Defaults to true.
    */
-  constructor(strategy, webapp = null, parameters = {}) {
-    super(strategy);
+  constructor(strategy, webapp = null, processors = [], parameters = {}) {
+    super(strategy, processors);
     this.output = process.stdout;
     const defaultParameters = {
       enableMethod: true,